+++ src/trx/game/objects/general/lights/on_off_light.c
@@ -7,7 +7,7 @@ static void M_Control(const int16_t item_num)
     if (!Item_IsTriggerActive(item)) {
         return;
     }
-    Output_AddDynamicLightRGB(item->pos, 16, (RGB_888) { 255, 255, 255 });
+    Output_AddDynamicLightRGB(item->pos, 16, COLOR_RGB_888_WHITE);
 }
 
 static void M_Setup(OBJECT *const obj)